requirements.txt
requests==2.24.0
Flask==1.1.2
pandas==1.0.5
Pillow==7.2.0
qrcode==6.1
